move paging out of extensions
